+++ src/main/java/swp/auctionkoi/service/payment/PaymentService.java
@@ -1,5 +1,5 @@
 package swp.auctionkoi.service.payment;
 
 public interface PaymentService {
-    public String requestWithdrawMoney(int userId, double amount) throws Exception;
+    public String requestWithdrawMoney(int userId, float amount) throws Exception;
 }
